+++ app/src/main/java/com/navapp/navigation/util/math/GraphLocation.java
@@ -1,5 +1,5 @@
 package com.navapp.navigation.util.math;
 
 public interface GraphLocation {
-    double distance(GraphLocation locations);
+    long distance(GraphLocation locations);
 }
